Extract parseResponse
